+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/TemplateMaintenance/ExistingTemplateContents.component.js
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import i18n from '@dhis2/d2-i18n';
 import { DialogTitle, DialogContent, DialogActions } from '@material-ui/core';
-import { Button } from '../../../Buttons/Button.component';
+import { Button } from '@dhis2/ui';
 
 const getStyles = () => ({
     buttonContainer: {