+++ src/core_modules/capture-ui/CoordinateField/CoordinateField.component.js
@@ -4,13 +4,12 @@ import classNames from 'classnames';
 import i18n from '@dhis2/d2-i18n';
 import { Map, TileLayer, Marker, withLeaflet } from 'react-leaflet';
 import { ReactLeafletSearch } from 'react-leaflet-search-unpolyfilled';
-import { IconCross24 } from '@dhis2/ui';
+import { IconCross24, Button } from '@dhis2/ui';
 import IconButton from '@material-ui/core/IconButton';
 import { AddLocationIcon } from '../Icons';
 import { CoordinateInput } from '../internal/CoordinateInput/CoordinateInput.component';
 import defaultClasses from './coordinateField.module.css';
 import { orientations } from '../constants/orientations.const';
-import { Button } from '../Buttons/Button.component';
 
 const WrappedLeafletSearch = withLeaflet(ReactLeafletSearch);